init redis connection, leaderboard adding score

+++ src/main/java/com/damian/leaderboardapi/dto/UpdateScoreDto.java
@@ -1,3 +1,3 @@
 package com.damian.leaderboardapi.dto;
 
-public record UpdateScoreDto(String userId, Integer score) { }
+public record UpdateScoreDto(String userId, Double score) { }
